Refactor so that old aclReachability ref still works
