+++ projects/question/src/main/java/org/batfish/question/AclReachabilityQuestionPlugin.java
@@ -8,7 +8,7 @@ import org.batfish.common.Answerer;
 import org.batfish.common.plugin.IBatfish;
 import org.batfish.common.plugin.Plugin;
 import org.batfish.datamodel.IpAccessList;
-import org.batfish.datamodel.answers.AclLinesOldAnswerElement;
+import org.batfish.datamodel.answers.AclLinesAnswerElement;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.collections.NamedStructureEquivalenceSets;
 import org.batfish.datamodel.questions.NodesSpecifier;
@@ -41,7 +41,7 @@ public class AclReachabilityQuestionPlugin extends QuestionPlugin {
       NamedStructureEquivalenceSets<?> aclEqSets =
           csnAnswer.getEquivalenceSets().get(IpAccessList.class.getSimpleName());
 
-      AclLinesOldAnswerElement answer = new AclLinesOldAnswerElement();
+      AclLinesAnswerElement answer = new AclLinesAnswerElement();
       _batfish.answerAclReachability(question.getAclNameRegex(), aclEqSets, answer);
       return answer;
     }
